fastapi/requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
python-jose==3.3.0
pydantic-settings==2.2.1
PyJWT==2.8.0